test: Move bedtools helper tests next to the helpers they cover
Three files under tests/unit/ tested helpers that live in
tests/integration/bedtools/utils/. Per the test guide, tests must
mirror the directory structure of the modules under test, and tests
for test infrastructure should sit alongside the infrastructure —
not in the project-level unit test package. The same three files
also wrapped module-level function tests in Test<Function> classes,
violating the guide's rule that module-level functions get
module-level tests.

Move the files:
  tests/unit/test_bedtools_wrapper.py -> tests/integration/bedtools/utils/
  tests/unit/test_comparison.py       -> tests/integration/bedtools/utils/
  tests/unit/test_duckdb_loader.py    -> tests/integration/bedtools/utils/

Convert TestCreateBedtool, TestIntersect, TestMerge, TestClosest,
TestBedtoolToTuples, TestCompareResults, and TestLoadIntervals from
class wrappers to module-level functions. TestBedtoolsError stays a
class because it wraps a real exception class. Update imports to
relative form and add pytestmark = pytest.mark.integration to each
moved module (these exercise real bedtools subprocess + DuckDB).
